tegra: hwpm: Modify OS common code for HWPM resmgr
- Modify the OS common code to be used by HWPM resource
manager in QNX.
- Add dev_name and fd fields in IP files
- Typecast variables to unsigned long long where ever they
are printed with %llx.

Jira THWPM-54

Reviewed-on: https://git-master.nvidia.com/r/c/linux-hwpm/+/2901186
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>

+++ drivers/tegra/hwpm/common/allowlist.c
@@ -122,8 +122,8 @@ int tegra_hwpm_combine_alist(struct tegra_soc_hwpm *hwpm, u64 *alist)
 	if (func_args.full_alist_idx != hwpm->alist_map->full_alist_size) {
 		tegra_hwpm_err(hwpm, "full_alist_size 0x%llx doesn't match "
 			"max full_alist_idx 0x%llx",
-			hwpm->alist_map->full_alist_size,
-			func_args.full_alist_idx);
+			(unsigned long long)hwpm->alist_map->full_alist_size,
+			(unsigned long long)func_args.full_alist_idx);
 		err = -EINVAL;
 	}
